bot-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
ccxt==4.3.56
pandas==2.2.2
pandas-ta==0.3.14b
apscheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cryptography==42.0.8
httpx==0.27.0
websockets==12.0
pydantic==2.7.4
sqlalchemy==2.0.31